careerpilot-ai/backend/requirements.txt
# requirements.txt
# CareerPilot AI Backend Final Production Version

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Data / ML
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Resume Parsing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Database
sqlalchemy>=2.0.0

# Authentication / Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation
pydantic>=2.0.0
email-validator>=2.0.0

# Utilities
aiofiles>=23.0.0
requests>=2.31.0

# Optional Production Helpers
gunicorn>=21.2.0